+++ src/main/java/com/wanted/socialintegratefreed/domain/auth/application/CustomUserDetailService.java
@@ -1,5 +1,6 @@
-package com.wanted.socialintegratefreed.domain.user.auth;
+package com.wanted.socialintegratefreed.domain.auth.application;
 
+import com.wanted.socialintegratefreed.domain.auth.entity.principal.UserPrincipal;
 import com.wanted.socialintegratefreed.domain.user.dao.UserRepository;
 import com.wanted.socialintegratefreed.domain.user.entity.User;
 import com.wanted.socialintegratefreed.global.error.BusinessException;
@@ -8,7 +9,6 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 
 /**
@@ -18,16 +18,17 @@ import org.springframework.stereotype.Service;
 @Service
 public class CustomUserDetailService implements UserDetailsService {
 
-    private final UserRepository userRepository;
+  private final UserRepository userRepository;
 
 
-    @Override
-    public UserDetails loadUserByUsername(String userEmail) throws UsernameNotFoundException {
-        User findUser =
-                userRepository.findByEmail(userEmail).orElseThrow(() -> new BusinessException("user email", userEmail,
-                        ErrorCode.EMAIL_NOT_EXIST));
-        // 유저를 찾은 이후 security 인증을위해 구현체 Principal 에 넣어줌
-        return new UserPrincipal(findUser);
-    }
+  @Override
+  public UserDetails loadUserByUsername(String userEmail) throws UsernameNotFoundException {
+    User findUser =
+        userRepository.findByEmail(userEmail)
+            .orElseThrow(() -> new BusinessException("user email", userEmail,
+                ErrorCode.EMAIL_NOT_EXIST));
+    // 유저를 찾은 이후 security 인증을위해 구현체 Principal 에 넣어줌
+    return new UserPrincipal(findUser);
+  }
 
 }
