+++ src/main/java/com/wanted/socialintegratefreed/domain/user/entity/UserEnable.java
@@ -1,0 +1,20 @@
+package com.wanted.socialintegratefreed.domain.user.entity;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+/**
+ * UserEnable: 로그인 여부
+ */
+public enum UserEnable {
+  /**
+   * @USER_ENABLED : 가입 여부 활성화
+   * @USER_DISABLED : 가입 여부 비활성화
+   */
+  USER_ENABLED("활성화"),
+  USER_DISABLED("비활성화");
+  private String enable;
+
+}
