+++ src/test/java/com/wanted/socialintegratefreed/domain/feed/api/FeedControllerTest.java
@@ -20,7 +20,7 @@ import com.wanted.socialintegratefreed.domain.feed.dto.response.FeedSearchRespon
 import com.wanted.socialintegratefreed.domain.feed.entity.Feed;
 import com.wanted.socialintegratefreed.domain.user.application.UserService;
 import com.wanted.socialintegratefreed.domain.user.entity.User;
-import com.wanted.socialintegratefreed.domain.user.jwt.JwtTokenProvider;
+import com.wanted.socialintegratefreed.global.config.security.jwt.JwtTokenProvider;
 import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
@@ -62,7 +62,7 @@ public class FeedControllerTest extends AbstractRestDocsTests {
   private Feed mockFeed;
 
   @BeforeEach
-  public void setUp(){
+  public void setUp() {
     mockMvc = MockMvcBuilders
         .webAppContextSetup(context)
         .apply(SecurityMockMvcConfigurers.springSecurity()) // 스프링 시큐리티 설정 적용
@@ -94,63 +94,63 @@ public class FeedControllerTest extends AbstractRestDocsTests {
         .andExpect(status().isCreated());
   }
 
-    @DisplayName("게시물 수정 요청 api가 성공한다.")
-    @Test
-    @WithMockUser(roles = {"USER"})
-    public void 게시물_수정() throws Exception {
-        Long feedId = 1L;
-        FeedUpdateRequest request = FeedUpdateRequest.builder()
-                .userId(1L)
-                .title("수정 제목")
-                .content("수정 내용")
-                .type(FeedType.FACEBOOK)
-                .build();
+  @DisplayName("게시물 수정 요청 api가 성공한다.")
+  @Test
+  @WithMockUser(roles = {"USER"})
+  public void 게시물_수정() throws Exception {
+    Long feedId = 1L;
+    FeedUpdateRequest request = FeedUpdateRequest.builder()
+        .userId(1L)
+        .title("수정 제목")
+        .content("수정 내용")
+        .type(FeedType.FACEBOOK)
+        .build();
 
-        mockMvc.perform(put("/api/v1/feeds/" + feedId)
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content(objectMapper.writeValueAsString(request)))
-                .andExpect(status().isOk());
-    }
-
-    @DisplayName("게시물 삭제 요청 api가 성공한다.")
-    @Test
-    @WithMockUser(roles = {"USER"})
-    public void 게시물_삭제() throws Exception {
-        Long feedId = 1L;
-
-        mockMvc.perform(delete("/api/v1/feeds/" + feedId))
-                .andExpect(status().isOk());
-    }
-
-    @DisplayName("게시물 조회 요청 api가 성공한다.")
-    @Test
-    @WithMockUser(roles = {"USER"})
-    public void 게시물_상세_조회() throws Exception {
-        Long feedId = 1L;
-
-        mockMvc.perform(get("/api/v1/feeds/" + feedId))
-                .andExpect(status().isOk());
-    }
-
-    @DisplayName("게시물 좋아요 api가 성공한다.")
-    @Test
-    @WithMockUser(roles = {"USER"})
-    void 게시물_좋아요() throws Exception {
-      Long feedId = 1L;
-
-      mockMvc.perform(post("/api/v1/feeds/" + feedId + "/like"))
+    mockMvc.perform(put("/api/v1/feeds/" + feedId)
+            .contentType(MediaType.APPLICATION_JSON)
+            .content(objectMapper.writeValueAsString(request)))
         .andExpect(status().isOk());
-    }
+  }
 
-    @DisplayName("게시물 공유 api가 성공한다.")
-    @Test
-    @WithMockUser(roles = {"USER"})
-    void 게시물_공유() throws Exception {
-      Long feedId = 1L;
+  @DisplayName("게시물 삭제 요청 api가 성공한다.")
+  @Test
+  @WithMockUser(roles = {"USER"})
+  public void 게시물_삭제() throws Exception {
+    Long feedId = 1L;
 
-      mockMvc.perform(post("/api/v1/feeds/" + feedId + "/share"))
+    mockMvc.perform(delete("/api/v1/feeds/" + feedId))
         .andExpect(status().isOk());
-    }
+  }
+
+  @DisplayName("게시물 조회 요청 api가 성공한다.")
+  @Test
+  @WithMockUser(roles = {"USER"})
+  public void 게시물_상세_조회() throws Exception {
+    Long feedId = 1L;
+
+    mockMvc.perform(get("/api/v1/feeds/" + feedId))
+        .andExpect(status().isOk());
+  }
+
+  @DisplayName("게시물 좋아요 api가 성공한다.")
+  @Test
+  @WithMockUser(roles = {"USER"})
+  void 게시물_좋아요() throws Exception {
+    Long feedId = 1L;
+
+    mockMvc.perform(post("/api/v1/feeds/" + feedId + "/like"))
+        .andExpect(status().isOk());
+  }
+
+  @DisplayName("게시물 공유 api가 성공한다.")
+  @Test
+  @WithMockUser(roles = {"USER"})
+  void 게시물_공유() throws Exception {
+    Long feedId = 1L;
+
+    mockMvc.perform(post("/api/v1/feeds/" + feedId + "/share"))
+        .andExpect(status().isOk());
+  }
 
   @DisplayName("게시물 통계 요청 api가 성공한다.")
   @Test
